pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claims-data-entry-agent"
version = "1.0.0"
description = "Automated claims data entry agent with OCR for TPA"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TPA Development Team"}
]
readme = "README.md"
keywords = ["claims", "ocr", "automation", "tpa", "paddleocr"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "aiofiles>=23.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "rq>=1.15.0",
    "paddlepaddle-gpu>=2.5.0",
    "numpy<2.0,>=1.23.0",
    "paddleocr==2.7.3",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0",
    "google-cloud-pubsub>=2.18.0",
    "gspread>=5.12.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "opencv-python<=4.6.0.66",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
